Google_colab
A classification model in digits dataset #classification: samples belong to two or more classes and to predict the class of unlabeled data. An example of a classification problem would be handwritten digit recognition, in which the aim is to assign each input vector to one of a finite number of discrete categories.